Passing envelopes#list tests

+++ db/migrations/20180824065138_create_envelopes.rb
@@ -3,7 +3,7 @@ Hanami::Model.migration do
     create_table :envelopes do
       primary_key :id
       column :name, String, null: false
-      column :amount, Float, null: false
+      column :amount, Integer, null: false
       column :created_at, DateTime, null: false
       column :updated_at, DateTime, null: false
     end
